fix(copy): integrated WriteLock on copy command
